Breaking up hearings actions (#10923)
