+++ client/app/hearings/components/ListSchedule.jsx
@@ -13,7 +13,7 @@ import {
   toggleTypeFilterVisibility, toggleLocationFilterVisibility,
   toggleVljFilterVisibility, onReceiveHearingSchedule,
   onViewStartDateChange, onViewEndDateChange
-} from '../actions/actions';
+} from '../actions/hearingScheduleActions';
 import { bindActionCreators } from 'redux';
 import connect from 'react-redux/es/connect/connect';
 import LoadingDataDisplay from '../../components/LoadingDataDisplay';
