+++ client/app/hearings/components/hearingWorksheet/HearingWorksheetStream.jsx
@@ -4,7 +4,7 @@ import { bindActionCreators } from 'redux';
 import _ from 'lodash';
 import Button from '../../../components/Button';
 import TabWindow from '../../../components/TabWindow';
-import { onAddIssue } from '../../actions/Issue';
+import { onAddIssue } from '../../actions/hearingWorksheetActions';
 import HearingWorksheetIssues from './HearingWorksheetIssues';
 
 class HearingWorksheetStream extends Component {
