+++ client/app/hearings/constants.js
@@ -6,7 +6,6 @@ export const ACTIONS = {
   RECEIVE_HEARING: 'RECEIVE_HEARING',
   RECEIVE_SAVED_HEARING: 'RECEIVE_SAVED_HEARING',
   RESET_SAVE_SUCCESSFUL: 'RESET_SAVE_SUCCESSFUL',
-  CANCEL_HEARING_UPDATE: 'CANCEL_HEARING_UPDATE',
   RECEIVE_HEARING_DAY_OPTIONS: 'RECEIVE_HEARING_DAY_OPTIONS',
   RECEIVE_UPCOMING_HEARING_DAYS: 'RECEIVE_UPCOMING_HEARING_DAYS',
   REGIONAL_OFFICE_CHANGE: 'REGIONAL_OFFICE_CHANGE',
@@ -60,7 +59,32 @@ export const ACTIONS = {
   RESET_DAILY_DOCKET_AFTER_SERVER_ERROR: 'RESET_DAILY_DOCKET_AFTER_SERVER_ERROR',
   HANDLE_LOCK_HEARING_SERVER_ERROR: 'HANDLE_LOCK_HEARING_SERVER_ERROR',
   RESET_LOCK_HEARING_SERVER_ERROR: 'RESET_LOCK_HEARING_SERVER_ERROR',
-  UPDATE_DOCKET_HEARING: 'UPDATE_DOCKET_HEARING'
+  UPDATE_DOCKET_HEARING: 'UPDATE_DOCKET_HEARING',
+  SET_REPNAME: 'SET_REPNAME',
+  SET_WITNESS: 'SET_WITNESS',
+  SET_HEARING_PREPPED: 'SET_HEARING_PREPPED',
+  POPULATE_WORKSHEET: 'POPULATE_WORKSHEET',
+  SET_HEARING_DAY_HEARINGS: 'SET_HEARING_DAY_HEARINGS',
+  SET_DESCRIPTION: 'SET_DESCRIPTION',
+  SET_ISSUE_NOTES: 'SET_ISSUE_NOTES',
+  SET_WORKSHEET_ISSUE_NOTES: 'SET_WORKSHEET_ISSUE_NOTES',
+  SET_ISSUE_DISPOSITION: 'SET_ISSUE_DISPOSITION',
+  SET_REOPEN: 'SET_REOPEN',
+  SET_ALLOW: 'SET_ALLOW',
+  SET_DENY: 'SET_DENY',
+  SET_REMAND: 'SET_REMAND',
+  SET_DISMISS: 'SET_DISMISS',
+  SET_OMO: 'SET_OMO',
+  ADD_ISSUE: 'ADD_ISSUE',
+  DELETE_ISSUE: 'DELETE_ISSUE',
+  TOGGLE_ISSUE_DELETE_MODAL: 'TOGGLE_ISSUE_DELETE_MODAL',
+  SET_MILITARY_SERVICE: 'SET_MILITARY_SERVICE',
+  SET_SUMMARY: 'SET_SUMMARY',
+  TOGGLE_WORKSHEET_SAVING: 'TOGGLE_WORKSHEET_SAVING',
+  SET_WORKSHEET_TIME_SAVED: 'SET_WORKSHEET_TIME_SAVED',
+  SET_ISSUE_EDITED_FLAG_TO_FALSE: 'SET_ISSUE_EDITED_FLAG_TO_FALSE',
+  SET_WORKSHEET_SAVE_FAILED_STATUS: 'SET_WORKSHEET_SAVE_FAILED_STATUS',
+  SET_WORKSHEET_EDITED_FLAG_TO_FALSE: 'SET_WORKSHEET_EDITED_FLAG_TO_FALSE'
 };
 
 export const ERROR_MAPPINGS = {
@@ -120,38 +144,6 @@ export const SPREADSHEET_TYPES = {
     template: '/JudgeAssignmentTemplate.xlsx'
   }
 };
-
-// actions
-//
-export const SET_REPNAME = 'SET_REPNAME';
-export const SET_WITNESS = 'SET_WITNESS';
-export const SET_HEARING_PREPPED = 'SET_HEARING_PREPPED';
-export const POPULATE_WORKSHEET = 'POPULATE_WORKSHEET';
-export const SET_HEARING_DAY_HEARINGS = 'SET_HEARING_DAY_HEARINGS';
-
-// issues
-export const SET_DESCRIPTION = 'SET_DESCRIPTION';
-export const SET_ISSUE_NOTES = 'SET_ISSUE_NOTES';
-export const SET_WORKSHEET_ISSUE_NOTES = 'SET_WORKSHEET_ISSUE_NOTES';
-export const SET_ISSUE_DISPOSITION = 'SET_ISSUE_DISPOSITION';
-export const SET_REOPEN = 'SET_REOPEN';
-export const SET_ALLOW = 'SET_ALLOW';
-export const SET_DENY = 'SET_DENY';
-export const SET_REMAND = 'SET_REMAND';
-export const SET_DISMISS = 'SET_DISMISS';
-export const SET_OMO = 'SET_OMO';
-export const ADD_ISSUE = 'ADD_ISSUE';
-export const DELETE_ISSUE = 'DELETE_ISSUE';
-export const TOGGLE_ISSUE_DELETE_MODAL = 'TOGGLE_ISSUE_DELETE_MODAL';
-
-export const SET_MILITARY_SERVICE = 'SET_MILITARY_SERVICE';
-export const SET_SUMMARY = 'SET_SUMMARY';
-
-export const TOGGLE_WORKSHEET_SAVING = 'TOGGLE_WORKSHEET_SAVING';
-export const SET_WORKSHEET_TIME_SAVED = 'SET_WORKSHEET_TIME_SAVED';
-export const SET_ISSUE_EDITED_FLAG_TO_FALSE = 'SET_ISSUE_EDITED_FLAG_TO_FALSE';
-export const SET_WORKSHEET_SAVE_FAILED_STATUS = 'SET_WORKSHEET_SAVE_FAILED_STATUS';
-export const SET_WORKSHEET_EDITED_FLAG_TO_FALSE = 'SET_WORKSHEET_EDITED_FLAG_TO_FALSE';
 
 export const DISPOSITION_OPTIONS = [{ value: 'held',
   label: 'Held' },
